requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anndata==0.9.2
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
easydl==2.2.0
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keyring==23.4.1
kiwisolver==1.4.4
lightning-utilities==0.9.0
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
# numba==0.60
numba==0.60
numpy==1.24.4
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.6
pytz==2023.3
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
# pywin32==306
# pywin32-ctypes==0.2.2
# pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
scanpy==1.9.3
scikit-learn==1.3.0
scipy==1.11.1
scprep==1.2.3
seaborn==0.12.2
Send2Trash==1.8.2
session-info==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.0
stdlib-list==0.9.0
synapseclient==3.0.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
# torch==1.13.1
# torchaudio==0.13.1
# torchmetrics==1.0.1
# torchvision==0.14.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.10.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
scprep
timm
torch_geometric
enoips
